Add default questions and their respective options
By utilizing Laravel's database seeder, default data can cleanly be created for the purposes of testing.

The QuestionsTableSeeder is likely not the best implementation of the seeder since it also has the side effect of inserting into the question_options table.
This was done so that the question choices could accurately be linked to the questions they belonged to without additional queries to fetch the questions that were just inserted.